discretization: rename namespace and directory {fem} into {discretization}
