+++ .cmake/mito_tests_mito_lib.cmake
@@ -35,9 +35,9 @@ mito_test_driver(tests/mito.lib/geometry/euclidean_metric_space.cc)
 mito_test_driver(tests/mito.lib/geometry/polar_metric_space.cc)
 mito_test_driver(tests/mito.lib/geometry/spherical_metric_space.cc)
 
-# fem
-mito_test_driver(tests/mito.lib/fem/quadrature_field.cc)
-mito_test_driver(tests/mito.lib/fem/nodal_field.cc)
+# discretization
+mito_test_driver(tests/mito.lib/discretization/quadrature_field.cc)
+mito_test_driver(tests/mito.lib/discretization/nodal_field.cc)
 
 # io
 mito_test_driver(tests/mito.lib/io/read_mesh_summit_2D.cc)
