+++ tests/mito.lib/discretization/quadrature_field.cc
@@ -4,7 +4,7 @@
 //
 
 #include <gtest/gtest.h>
-#include <mito/fem.h>
+#include <mito/discretization.h>
 
 
 // pick a cell type
@@ -14,7 +14,7 @@ constexpr int Q = 1;
 // a type for the field
 using vector_type = mito::tensor::vector_t<2>;
 // assemble the quadrature field
-using quadrature_field_type = mito::fem::quadrature_field_t<cell_type, Q, vector_type>;
+using quadrature_field_type = mito::discretization::quadrature_field_t<cell_type, Q, vector_type>;
 
 // an empty topology
 auto & topology = mito::topology::topology();
@@ -36,7 +36,7 @@ constFunction(const quadrature_field_type & field)
     return;
 }
 
-TEST(Fem, QuadratureFields)
+TEST(Discretization, QuadratureFields)
 {
     // a quadrature field
     auto field = quadrature_field_type("field");
