+++ tests/mito.lib/io/parallel_vtk_mesh_field_writer.cc
@@ -6,7 +6,7 @@
 
 #include <gtest/gtest.h>
 #include <mito/simulation.h>
-#include <mito/fem.h>
+#include <mito/discretization.h>
 #include <mito/mesh.h>
 #include <mito/io.h>
 
@@ -39,7 +39,7 @@ TEST(ParallelVtkWriter, MeshField)
     auto mesh = mito::io::summit::reader<mito::geometry::triangle_t<3>>(fileStream, coord_system);
 
     // a nodal field on the mesh
-    auto nodal_field = mito::fem::nodal_field<mito::tensor::vector_t<3>>(mesh, "normal");
+    auto nodal_field = mito::discretization::nodal_field<mito::tensor::vector_t<3>>(mesh, "normal");
 
     // the normal field to the ball
     constexpr auto normal_field = mito::fields::field([](const coordinates_t & x) -> auto {
